pyproject.toml
[project]
name = "kubeflow-mcp"
version = "0.1.0-dev"
requires-python = ">=3.10"
authors = [
    {name = "Abhijeet Dhumal", email = "abhijeetdhumal652@gmail.com"},
    {name = "The Kubeflow Authors", email = "kubeflow-discuss@googlegroups.com"},
]
license = "Apache-2.0"
license-files = ["LICENSE"]
description = "AI-powered interface for Kubeflow Training via Model Context Protocol (MCP)"
readme = "README.md"
keywords = ["kubeflow", "mcp", "model-context-protocol", "trainer", "llm", "ai", "fine-tuning", "distributed-training"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "fastmcp>=2.0.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "click>=8.0",
    "kubernetes>=29.0",
]

[project.optional-dependencies]
trainer = ["kubeflow>=0.4.0", "huggingface_hub>=0.20"]
optimizer = ["kubeflow-katib>=0.17.0"]
hub = ["model-registry>=0.2.0"]
all = [
    "kubeflow-training>=1.8.0",
    "kubeflow-katib>=0.17.0",
    "model-registry>=0.2.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "types-PyYAML>=6.0",
]
benchmark = [
    "pytest-benchmark>=4.0",
    "tiktoken>=0.7",
    "matplotlib>=3.8",
]
docs = [
    "sphinx>=7.0",
    "sphinx-rtd-theme>=2.0",
]
agents = [
    "llama-index-core>=0.12",
    "llama-index-llms-ollama>=0.4",
    "rich>=13.0",
]

[project.scripts]
kubeflow-mcp = "kubeflow_mcp.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/kubeflow_mcp"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "UP",   # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["kubeflow_mcp"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
